+++ ConsoleExample/Models/ExampleConstants.cs
@@ -11,11 +11,7 @@ namespace ConsoleExample.Models
         public const string BUTTONS_FILE_KEY = "buttons";
         public const string MESSAGES_FILE_KEY = "messages";
 
-
         public const string MSG_MAIN_MENU   = "";
-
-        
-
 
         public const string DYNAMIC_COMMANT_EXAMPLE = "";